Api/requirements.txt
fastapi
sqlalchemy
uvicorn
pytz
python-jose
passlib
python-multipart
pyjwt
bcrypt
psycopg2-binary
psycopg2
python-dotenv
pydantic[email]
aiofiles